Config/requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
cryptography==43.0.3
dj-user-login-history==1.0.6
Django==5.1.2
django-admin-logs==1.3.0
django-appconf==1.0.6
django-browser-reload==1.17.0
django-ckeditor==6.7.1
django-debug-toolbar==4.4.6
django-defender==0.9.8
django-extensions==3.2.3
django-hijack==3.7.0
django-htmx==1.21.0
django-imagekit==5.0.0
django-jazzmin==3.0.1
django-js-asset==2.2.0
django-request==1.6.3
django-tailwind==3.8.0
Faker==30.8.2
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pilkit==3.0
pillow==11.0.0
pycparser==2.22
Pygments==2.18.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
rich==13.9.4
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.1
